Factor out boolean engine and improve selection/testing
